study_notes_ai/pyproject.toml
[project]
name = "study-notes-ai"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "gemini-cli>=0.2.5",
    "openai-agents>=0.6.1",
    "pypdf2>=3.0.1",
    "python-dotenv>=1.2.1",
    "streamlit>=1.51.0",
]
